TwitchPlaysLeaderboard update
This update solely deals with the website twitchplaysleaderboard.info, which collects track information from this list. Some items were not coming through because they were not spaced correctly - others had comments in them that I had no idea were parsed. This fixes both.

There is no change that needs to be done to PBR itself yet. The next refresh can be reserved for the next update, which can probably be in a month or two.