The Totino's Fix
Unfortunately, during the development of Music Update 80, something which happened which caused the Totino's Mania music to not play. We are currently in the process of troubleshooting. Please understand.

In the meantime, please take some extra music as a consolation. It's completely original stuff, and not something I was preparing for music update 81